Renamed 'tasks' to 'scripts' and moved in 'utils'
