+++ scripts/utils/ansi-colors.js
