+++ scripts/utils/license-header.js
